backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.15.2,<3.16
Pillow>=11.1.0,<11.2.0
drf-spectacular>=0.28.0,<0.29.0
psycopg2>=2.9.10,<2.10
django-cors-headers
scikit-learn==1.2.2
pandas>=2.2.2,<2.3
numpy==1.24.4
